+++ BTCPayServer.Abstractions/Custodians/Client/Exception/InsufficientFundsException.cs
@@ -1,0 +1,9 @@
+namespace BTCPayServer.Abstractions.Custodians;
+
+public class InsufficientFundsException : CustodianApiException
+{   
+    public InsufficientFundsException(string message) : base(400, "insufficient-funds", message)
+    {
+    }
+
+}