+++ BTCPayServer.Client/Models/CustodianAccountResponse.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+
+namespace BTCPayServer.Client.Models;
+
+public class CustodianAccountResponse: CustodianAccountData
+{
+    public IDictionary<string, decimal> AssetBalances { get; set; }
+
+    public CustodianAccountResponse()
+    {
+        
+    }
+
+}