+++ BTCPayServer/Controllers/GreenField/GreenfieldLightningNodeApiController.cs
@@ -6,10 +6,8 @@ using BTCPayServer.Abstractions.Contracts;
 using BTCPayServer.Abstractions.Extensions;
 using BTCPayServer.Client;
 using BTCPayServer.Client.Models;
-using BTCPayServer.HostedServices;
 using BTCPayServer.Lightning;
 using BTCPayServer.Security;
-using BTCPayServer.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
@@ -50,7 +48,7 @@ namespace BTCPayServer.Controllers.Greenfield
             });
         }
 
-        public virtual async Task<IActionResult> ConnectToNode(string cryptoCode, ConnectToNodeRequest request)
+        public virtual async Task<IActionResult> ConnectToNode(string cryptoCode, ConnectToNodeRequest request, CancellationToken cancellationToken = default)
         {
             var lightningClient = await GetLightningClient(cryptoCode, true);
             if (request?.NodeURI is null)
@@ -63,7 +61,7 @@ namespace BTCPayServer.Controllers.Greenfield
                 return this.CreateValidationError(ModelState);
             }
 
-            var result = await lightningClient.ConnectTo(request.NodeURI);
+            var result = await lightningClient.ConnectTo(request.NodeURI, cancellationToken);
             switch (result)
             {
                 case ConnectionResult.Ok:
@@ -158,10 +156,10 @@ namespace BTCPayServer.Controllers.Greenfield
             return this.CreateAPIError(errorCode, errorMessage);
         }
 
-        public virtual async Task<IActionResult> GetDepositAddress(string cryptoCode)
+        public virtual async Task<IActionResult> GetDepositAddress(string cryptoCode, CancellationToken cancellationToken = default)
         {
             var lightningClient = await GetLightningClient(cryptoCode, true);
-            return Ok(new JValue((await lightningClient.GetDepositAddress()).ToString()));
+            return Ok(new JValue((await lightningClient.GetDepositAddress(cancellationToken)).ToString()));
         }
 
         public virtual async Task<IActionResult> GetPayment(string cryptoCode, string paymentHash, CancellationToken cancellationToken = default)
@@ -187,8 +185,8 @@ namespace BTCPayServer.Controllers.Greenfield
                 return this.CreateValidationError(ModelState);
             }
             
-            var param = lightningInvoice?.MaxFeeFlat != null || lightningInvoice?.MaxFeePercent != null
-                ? new PayInvoiceParams { MaxFeePercent = lightningInvoice.MaxFeePercent, MaxFeeFlat = lightningInvoice.MaxFeeFlat }
+            var param = lightningInvoice?.MaxFeeFlat != null || lightningInvoice?.MaxFeePercent != null || lightningInvoice?.Amount != null
+                ? new PayInvoiceParams { MaxFeePercent = lightningInvoice.MaxFeePercent, MaxFeeFlat = lightningInvoice.MaxFeeFlat, Amount = lightningInvoice.Amount }
                 : null;
             var result = await lightningClient.Pay(lightningInvoice.BOLT11, param, cancellationToken);