+++ BTCPayServer/PayoutProcessors/Lightning/UILightningAutomatedPayoutProcessorsController.cs
@@ -101,10 +101,10 @@ public class UILightningAutomatedPayoutProcessorsController : Controller
             Id = activeProcessor.Id,
             Processed = tcs
         });
-        TempData.SetStatusMessageModel(new StatusMessageModel()
+        TempData.SetStatusMessageModel(new StatusMessageModel
         {
             Severity = StatusMessageModel.StatusSeverity.Success,
-            Message = $"Processor updated."
+            Message = "Processor updated."
         });
         await tcs.Task;
         return RedirectToAction("ConfigureStorePayoutProcessors", "UiPayoutProcessors", new {storeId});
@@ -121,11 +121,12 @@ public class UILightningAutomatedPayoutProcessorsController : Controller
         {
             IntervalMinutes = blob.Interval.TotalMinutes;
         }
+        
         public double IntervalMinutes { get; set; }
 
         public AutomatedPayoutBlob ToBlob()
         {
-            return new AutomatedPayoutBlob() { Interval = TimeSpan.FromMinutes(IntervalMinutes) };
+            return new AutomatedPayoutBlob { Interval = TimeSpan.FromMinutes(IntervalMinutes) };
         }
     }
 }